requirements.txt
streamlit
openai
python-dotenv
pandas
sqlparse
llama-cpp-python
